% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_model.R
\name{delete_model}
\alias{delete_model}
\title{Unload loaded model}
\usage{
delete_model(model_name, req, res)
}
\arguments{
\item{model_name}{Name of the model}

\item{req}{Request object}

\item{res}{Response object}
}
\value{
Unloads a loaded model
}
\description{
Unload loaded model
}
