% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Information from loaded model}
\usage{
get_model(model_name, req, res)
}
\arguments{
\item{model_name}{Name of the model}

\item{req}{Request object}

\item{res}{Response object}
}
\value{
Gets information from a loaded model
}
\description{
Information from loaded model
}
