% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke_model.R
\name{invoke_model}
\alias{invoke_model}
\title{Title}
\usage{
invoke_model(model_name, data_for_inference, req, res)
}
\arguments{
\item{model_name}{Name of the model}

\item{data_for_inference}{The data needed for the model}

\item{req}{Request object}

\item{res}{Response object}
}
\value{
A prediction from `model_name`.
}
\description{
Title
}
