% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_model.R
\name{list_model}
\alias{list_model}
\title{Title}
\usage{
list_model(req, res)
}
\arguments{
\item{req}{Request object}

\item{res}{Response object}
}
\value{
List of loaded models tracked in `loaded_models`
}
\description{
Title
}
