% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invocations.R
\name{preprocess}
\alias{preprocess}
\title{Single model endpoint}
\usage{
preprocess(raw_data, req, res)
}
\arguments{
\item{raw_data}{Unprocessed data}

\item{req}{Request object}

\item{res}{Response object}
}
\value{
Gets information from a loaded model
}
\description{
Single model endpoint
}
